Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
mysqlclient
python-dotenv
passlib
python-jose
numpy
pandas
plotly==5.15.0
scipy==1.14.0
snowflake-connector-python==3.12.2
snowflake-snowpark-python==1.9.0
scipy
snowflake==0.7.0
openpyxl
xlrd
requests
msal
bcrypt
joblib
psutil
redis
aioredis